requirements.txt
appier==1.31.5
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
click==8.1.7
google-api-core==2.15.0
google-api-python-client==2.2.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
importmagic3==0.2.0
jedi==0.18.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
Pillow==8.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.2.8
PyJWT==2.8.0
pyparsing==2.4.7
python-magic==0.4.27
pytz==2021.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
spark-parser==1.8.9
twilio==8.11.1
uncompyle6==3.7.4
uritemplate==4.1.1
urllib3==2.1.0
wdb==3.3.0
xdis==5.0.9
